Refactor and add Unit tests
