+++ TODO.txt
@@ -1,0 +1,10 @@
+Proper error level handling for label Rules
+Act on 'count' in required rules.
+Add count to optional rules.
+Refactor code into small separate files.
+
+Automated test integration
+Comment samples
+Update README
+
+Add jslinting
