+++ test/unit/rulefile-loader.spec.js
@@ -1,0 +1,7 @@
+'use strict';
+
+require('should');
+
+
+// it('Should allow underscores', function() {;
+// });
